+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [0.6.1] - 2022-02-25
 
 ### Added
 - Add the `none` value to the json DSL
@@ -167,7 +167,8 @@ Was yanked and replaced by [0.5.0]
 - async `UserFunction` support
 
 
-[unreleased]: https://github.com/mendelt/reval/compare/v0.6.0...HEAD
+[unreleased]: https://github.com/mendelt/reval/compare/v0.6.1...HEAD
+[0.6.1]: https://github.com/mendelt/reval/compare/v0.6.0...v0.6.1
 [0.6.0]: https://github.com/mendelt/reval/compare/v0.5.2...v0.6.0
 [0.5.2]: https://github.com/mendelt/reval/compare/v0.5.1...v0.5.2
 [0.5.1]: https://github.com/mendelt/reval/compare/v0.5.0...v0.5.1